#37287 #37353 (2.2.0.89) 型の置換を継続中。 / Ongoing type replacement.

+++ src/wizard1.c
@@ -2216,7 +2216,7 @@ static void spoil_mon_info(cptr fname)
 		spoil_out(buf);
 
 		/* Level */
-		sprintf(buf, "Lev:%d  ", r_ptr->level);
+		sprintf(buf, "Lev:%d  ", (int)r_ptr->level);
 		spoil_out(buf);
 
 		/* Rarity */
